+++ royal/tests/functional/test_sa_example.py
@@ -46,15 +46,31 @@ class Test(TestBase):
 
     def test_project_update(self):
         self.app.patch('/projects/999', {'name': 'new_name'}, status=404)
-        self.app.patch('/projects/%s' % PR_ID, {'name': 'new_name'})
+        result = self.app.patch('/projects/%s' % PR_ID, {'name': 'new_name'})
 
         expected = {'id': PR_ID,
                     'links': {'self': 'http://localhost/projects/%s/' % PR_ID},
                     'name': u'new_name'}
-        result = self.app.get('/projects/%s' % PR_ID).json
-        self.assertEqual(result, expected)
+        self.assertEqual(result.json, expected)
 
     def test_project_create_duplicate(self):
 
         with self.assertRaises(IntegrityError):
             self.app.post('/projects', {'name': PROJECT_NAME})
+
+    def test_project_replace_existing(self):
+        result = self.app.put('/projects/%s' % PR_ID, {'name': 'new_name'})
+        expected = {'id': PR_ID,
+                    'links': {'self': 'http://localhost/projects/%s/' % PR_ID},
+                    'name': u'new_name'}
+        self.assertEqual(result.json, expected)
+
+    def test_project_replace_inexisting(self):
+        from example.ext.sqla.model import Project
+        result = self.app.put('/projects/999', {'name': 'new_name'})
+        expected = {'id': '999',  # XXX Fixme: must be an integer
+                    'links': {'self': 'http://localhost/projects/999/'},
+                    'name': u'new_name'}
+        self.assertEqual(result.json, expected)
+
+        self.assertIsNotNone(self.session.query(Project).get(999))